Added prerendering for canvas sprites. Runs when the sprite is first created and when the resolution is changed. Called with the same arguments as the render function. It can also be called manually by using sprite.prerender(), this actually calls a wrapper function so you don't need to provide any arguments.
Canvas render and prerender functions are now called with scaleX and 
scaleY as their last two arguments.

The textures created for canvas sprites are now deleted when the canvas 
sprite is deleted.

Added a new argument to Game.playSound: volume. This is the 2nd argument 
so the loop argument is now the 3rd. The 4th argument is "restart" which 
is if the sound should restart if it's already playing, defaults to 
false.

Fixed the touching.sprite method so it now works properly with clones.

Layer operations are now cancelled if the sprite is deleted before 
they're run. Prevents major graphical glitches.

Fixed a bug where clicks would be registered when they weren't on the 
canvas.

The right click context menu is now hidden.

 Bagel.internal.getTypeOf now reports NaN as type "NaN" instead of 
"number".

Bagel.internal.deepClone now won't clone objects have the property 
internal.dontClone set to true. (keeps it as a reference)

Fixed sprite.last for canvas sprites.

Changed the static canvas demo so it now uses the prerender method.