Changed how text sprites work: The font argument is now just the font instead of a canvas font. The size argument is now used to set the font size (still in game pixels).
Added bitmap fonts and texts. Create the fonts by putting characters in 
a vertical row in an image and use red to mark the start and early end 
if applicable. See assets/imgs/pixelFont.png as an example. Once you've 
created it, make it into a font using Bagel.font.create and selecting 
the image.

To use the font in a text sprite, load the font in assets, set the 
"bitmap" argument to true and set the font to the id.

Added a new test/example for bitmap text.

Started adding a basic compression thing, currently doesn't work 
properly. Will probably use it for fonts when it's working.