Added webp and ogg support. It was supported before but now fallback is supported.
Set the webp/ogg argument of an image/sound asset and it will be used if 
it's supported by the browser, otherwise the src is used.

Corrected the built in plugin description as the renderer sprite type 
was removed ages ago.

Improved the code for the default sprite type's listeners. Currently not 
finished apparently?

Changed src to href in some descriptions.

Added a new argument for asset types: "hrefArgs", it's an array 
containing the arguments that are hrefs for generating service workers. 
Specifiying the "src" argument is optional.

Added plugin proxies (Bagel.internal.current.pluginProxy) which when set 
to true, means the current plugin is acting on behalf of the previous. 
This means that you can't use the .Internal prefix and you can actually 
use that plugin's reserved prefix. Can find out which plugin is acting 
by using Bagel.internal.getActingPluginId(). You can also specify true 
as the first argument to say that this plugin is a proxy and it's set 
back after.

Started adding steps for clones. Can't remember if I finished.

Bagel.internal.deepClone now ignores the dontClone argument for the 
first dimension.

Tidied up the code for saving current.

WebP support is now detected in Firefox via a special case despite it 
not having encoding support.