Added word wrapping to both text types. To enable, set "wordWrapWidth" to the maximum width of the text sprite in game pixels.
Changing text modes now correctly defaults to a valid font for that 
type.

The other unmute button image is now preloaded on creation.

Removed initial delay logic and references.

Added a word wrap demo.

Tidied up the multi line text demo.

The default font is now bundled in with Bagel.js. 
(.Internal.defaultFont, but it's the default if unspecified)

Readded setting assets in a more restrictive way: setting the "set" 
property of an asset type in the plugin allows it and splits the asset 
loading into two steps. "init" will run when the asset is loaded 
normally, but the ready function is replaced with a function to call the 
set function. It's called with the content of the network request.

The set function is then called with these arguments: asset, assetJSON, 
ready, game, plugin. Where asset is the result of the network request 
and ready behaves as normal. Call ready when it's done (usually 
synchronously). If an asset is set, only the set function will be run 
(in the same way). So make sure to check the content.

Fixed a "sprite" sprite bug apparently.

Canvases now only prerender when they are made visible.