Added sprite and bitmap sprite cropping and improved sound network load.
Both sprites (the default type) and bitmapSprites can be cropped with 
the crop argument.

If the duration of an audio asset is specified in the asset JSON, the 
sound won't be preloaded to reduce network usage. The sound will still 
be marked as loaded though and will have the specified duration.